improvement: Merge Container Summary view with summary header
