+++ packages/tools/client-debugger/client-debugger-view/src/components/index.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-// TODO: re-enable rule and fix violations before merging into main
+// TODO: re-enable rule and fix violations
 /* eslint-disable no-restricted-syntax */
 
 export * from "./client-data-views";
@@ -11,7 +11,6 @@ export * from "./data-object-views";
 
 export * from "./AudienceView";
 export * from "./ClientDebugView";
-export * from "./ContainerDataView";
 export * from "./ContainerSelectionDropdown";
 export * from "./ContainerStateView";
