+++ packages/tools/client-debugger/client-debugger-view/src/index.ts
@@ -21,7 +21,7 @@
  */
 
 export {
-	AudienceMemberViewProps,
+	AudienceMemberViewProps, // TODO: remove this
 	ClientDebugView,
 	clientDebugViewClassName,
 	ClientDebugViewProps,
